+++ python/tests/sync_tests/test_sync_pubsub.py
@@ -4,7 +4,7 @@ from __future__ import annotations
 
 import time
 from enum import IntEnum
-from typing import Any, Dict, List, Optional, Set, Tuple, Union, cast
+from typing import Any, List, Optional, Tuple, Union, cast
 
 import pytest
 from glide_shared.commands.core_options import PubSubMsg
@@ -19,7 +19,10 @@ from glide_sync.glide_client import GlideClient, GlideClusterClient, TGlideClien
 
 from tests.sync_tests.conftest import create_sync_client
 from tests.utils.utils import (
+    create_pubsub_subscription,
+    decode_pubsub_msg,
     get_random_string,
+    new_message,
     run_sync_func_with_timeout_in_thread,
     sync_check_if_server_version_lt,
 )
@@ -89,16 +92,6 @@ def create_two_clients_with_pubsub(
     return client1, client2
 
 
-def decode_pubsub_msg(msg: Optional[PubSubMsg]) -> PubSubMsg:
-    if not msg:
-        return PubSubMsg("", "", None)
-    string_msg = cast(bytes, msg.message).decode()
-    string_channel = cast(bytes, msg.channel).decode()
-    string_pattern = cast(bytes, msg.pattern).decode() if msg.pattern else None
-    decoded_msg = PubSubMsg(string_msg, string_channel, string_pattern)
-    return decoded_msg
-
-
 def get_message_by_method(
     method: MethodTesting,
     client: TGlideClient,
@@ -131,35 +124,6 @@ def check_no_messages_left(
     else:
         assert callback is not None
         assert len(callback) == expected_callback_messages_count
-
-
-def create_pubsub_subscription(
-    cluster_mode,
-    cluster_channels_and_patterns: Dict[
-        GlideClusterClientConfiguration.PubSubChannelModes, Set[str]
-    ],
-    standalone_channels_and_patterns: Dict[
-        GlideClientConfiguration.PubSubChannelModes, Set[str]
-    ],
-    callback=None,
-    context=None,
-):
-    if cluster_mode:
-        return GlideClusterClientConfiguration.PubSubSubscriptions(
-            channels_and_patterns=cluster_channels_and_patterns,
-            callback=callback,
-            context=context,
-        )
-    return GlideClientConfiguration.PubSubSubscriptions(
-        channels_and_patterns=standalone_channels_and_patterns,
-        callback=callback,
-        context=context,
-    )
-
-
-def new_message(msg: PubSubMsg, context: Any):
-    received_messages: List[PubSubMsg] = context
-    received_messages.append(msg)
 
 
 def client_cleanup(
